+++ tests/tests_settings/test_settings_files.py
@@ -47,20 +47,26 @@ def test_wl_table_tags():
     table_tags.defaults_row = ['test', 'test', 'test', '']
     table_tags._add_row(row = 0)
 
+    table_tags.model().item(0, 2).setText('')
+    table_tags.check_empty_duplicate_tags(table_tags.model().item(0, 2))
+
     table_tags.reset_table()
     table_tags.get_tags()
 
 def test_wl_table_tags_header():
     table_tags_header = wl_settings_files.Wl_Table_Tags_Header(main)
-    table_tags_header.item_changed(table_tags_header.model().item(0, 0))
+    table_tags_header.model().item(0, 2).setText('')
+    table_tags_header.item_changed(table_tags_header.model().item(0, 2))
 
 def test_wl_table_tags_body():
     table_tags_body = wl_settings_files.Wl_Table_Tags_Body(main)
-    table_tags_body.item_changed(table_tags_body.model().item(0, 0))
+    table_tags_body.model().item(0, 2).setText('')
+    table_tags_body.item_changed(table_tags_body.model().item(0, 2))
 
 def test_wl_table_tags_xml():
     table_tags_xml = wl_settings_files.Wl_Table_Tags_Xml(main)
-    table_tags_xml.item_changed(table_tags_xml.model().item(0, 0))
+    table_tags_xml.model().item(0, 2).setText('')
+    table_tags_xml.item_changed(table_tags_xml.model().item(0, 2))
 
 if __name__ == '__main__':
     test_wl_settings_files()