+++ tests/tests_settings/test_settings_lemmatization.py
@@ -20,6 +20,7 @@ from tests import (
     wl_test_init,
     wl_test_lang_examples
 )
+from wordless.wl_nlp import wl_nlp_utils
 from wordless.wl_settings import wl_settings_lemmatization
 
 main = wl_test_init.Wl_Test_Main(switch_lang_utils = 'fast')
@@ -38,6 +39,12 @@ def test_wl_settings_lemmatization():
     main.settings_lemmatization.preview_results_changed()
     main.settings_lemmatization.update_gui('test')
     main.settings_lemmatization.update_gui_err()
+
+    # Force the model download to fail
+    worker_download_model_stanza_temp = wl_nlp_utils.Wl_Worker_Download_Model_Stanza
+    wl_nlp_utils.Wl_Worker_Download_Model_Stanza = wl_test_init.Wl_Worker_Download_Model_Stanza_Err_Msg
+    main.settings_lemmatization.preview_results_changed()
+    wl_nlp_utils.Wl_Worker_Download_Model_Stanza = worker_download_model_stanza_temp
 
 def test_wl_worker_preview_lemmatizer():
     main.settings_custom['lemmatization']['preview']['preview_samples'] = wl_test_lang_examples.TEXT_NEWLINES