+++ tests/tests_settings/test_settings_sentence_tokenization.py
@@ -40,6 +40,12 @@ def test_wl_settings_sentence_tokenization():
     main.settings_sentence_tokenization.update_gui('test')
     main.settings_sentence_tokenization.update_gui_err()
 
+    # Force the model download to fail
+    worker_download_model_stanza_temp = wl_nlp_utils.Wl_Worker_Download_Model_Stanza
+    wl_nlp_utils.Wl_Worker_Download_Model_Stanza = wl_test_init.Wl_Worker_Download_Model_Stanza_Err_Msg
+    main.settings_sentence_tokenization.preview_results_changed()
+    wl_nlp_utils.Wl_Worker_Download_Model_Stanza = worker_download_model_stanza_temp
+
 def test_wl_worker_preview_sentence_tokenizer():
     main.settings_custom['sentence_tokenization']['preview']['preview_samples'] = wl_test_lang_examples.TEXT_NEWLINES
     preview_lang = main.settings_custom['sentence_tokenization']['preview']['preview_lang']